feat: refactorgin soap encoding
