+++ src/main/java/it/pagopa/transactions/utils/soap/Jaxb2SoapEncoder.java
@@ -1,6 +1,7 @@
 package it.pagopa.transactions.utils.soap;
 
 import org.reactivestreams.Publisher;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.ResolvableType;
 import org.springframework.core.codec.CodecException;
 import org.springframework.core.codec.Encoder;
@@ -9,11 +10,11 @@ import org.springframework.core.io.buffer.DataBuffer;
 import org.springframework.core.io.buffer.DataBufferFactory;
 import org.springframework.core.io.buffer.DataBufferUtils;
 import org.springframework.core.io.buffer.PooledDataBuffer;
-import org.springframework.util.ClassUtils;
 import org.springframework.util.MimeType;
 import org.springframework.util.MimeTypeUtils;
 import org.w3c.dom.Document;
 
+import it.pagopa.transactions.model.SoapEnvelope;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
@@ -41,7 +42,6 @@ public class Jaxb2SoapEncoder implements Encoder<Object> {
         Class<?> outputClass = elementType.toClass();
         return (outputClass.isAnnotationPresent(XmlRootElement.class) ||
                 outputClass.isAnnotationPresent(XmlType.class));
-
     }
 
     @Override
@@ -65,33 +65,36 @@ public class Jaxb2SoapEncoder implements Encoder<Object> {
             Map<String, Object> hints) {
 
         return Mono.fromCallable(() -> {
+
             boolean release = true;
             DataBuffer buffer = bufferFactory.allocateBuffer(1024);
             DocumentBuilderFactory dbf = DocumentBuilderFactory.newInstance();
+            SoapEnvelope soapEnvelope = (SoapEnvelope) value;
+            OutputStream outputStream = buffer.asOutputStream();
+            Document doc = dbf.newDocumentBuilder().newDocument();
+            SOAPMessage soapMessage = MessageFactory.newInstance().createMessage();
 
             try {
 
-                Document doc = dbf.newDocumentBuilder().newDocument();
+                /**
+                 * Marshal: body content into a xml Document.
+                 */
+                getMarshaller().marshal(soapEnvelope.getBody(), doc);
 
-                SoapEnvelopeRequest soapEnvelopeRequest = (SoapEnvelopeRequest) value;
-
-                OutputStream outputStream = buffer.asOutputStream();
-                Class<?> clazz = ClassUtils.getUserClass(soapEnvelopeRequest.getBody());
-                Marshaller marshaller = initMarshaller(clazz);
-
-                marshaller.marshal(soapEnvelopeRequest.getBody(), doc);
-
-                SOAPMessage soapMessage = MessageFactory.newInstance().createMessage();
-
+                /**
+                 * Clean: SOAP-ENV (SOAP1.2) into soap (SOAP1.1)
+                 */
                 soapMessage.getSOAPPart().getEnvelope().removeNamespaceDeclaration("SOAP-ENV");
                 soapMessage.getSOAPPart().getEnvelope().setPrefix("soap");
                 soapMessage.getSOAPHeader().setPrefix("soap");
                 soapMessage.getSOAPBody().setPrefix("soap");
 
+                /**
+                 * Marshal ending
+                 */
                 SOAPBody soapBody = soapMessage.getSOAPBody();
                 soapBody.addDocument(doc);
                 soapMessage.saveChanges();
-
                 soapMessage.writeTo(outputStream);
                 outputStream.flush();
 
@@ -110,10 +113,9 @@ public class Jaxb2SoapEncoder implements Encoder<Object> {
         }).flux();
     }
 
-    private Marshaller initMarshaller(Class<?> clazz) throws JAXBException {
-        Marshaller marshaller = this.jaxbContexts.createMarshaller(clazz);
+    private Marshaller getMarshaller() throws JAXBException {
+        Marshaller marshaller = this.jaxbContexts.createMarshaller();
         marshaller.setProperty(Marshaller.JAXB_ENCODING, StandardCharsets.UTF_8.name());
         return marshaller;
     }
-
 }
